Begin Split
Contestant has to be split into model and controller components. This commit is saving progress.

+++ Assets/Resources/Scripts/Poker/Commands/RaiseCommand.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class RaiseCommandArgs
 {
     public Contestant Contestant;
-    public int RaisedBet;
+    public long RaisedBet;
     public PokerType Game;
 
-    public RaiseCommandArgs(Contestant c, int r, PokerType g)
+    public RaiseCommandArgs(Contestant c, long r, PokerType g)
     {
         Contestant = c;
         RaisedBet = r;
